issue 35503: image/gif: decode fails with "gif: too much image data"
The author provides three example images which fail to decode
with the standard library's 'image/gif' package. The underlying
cause appears to be reading the "End of Information Code" from
the LZW decoder.

The decoder returns 1 byte read instead of zero, as well as a nil
error instead of io.EOF or io.ErrUnexpectedEOF. This triggers
the errTooMuch condition.

The solution to this is to comment out the errTooMuch return,
this results in a successful decode, but also fails some of the
tests. More investigation is necessary to determine whether
this solution is sound in practice.